move client and schema out of sdk subpackage
