+++ aito/cli/sub_commands/infer_table_schema_sub_command.py
@@ -3,8 +3,8 @@ import sys
 from typing import Dict
 from typing import List
 
-from aito.sdk.aito_schema import AitoTableSchema
-from aito.sdk.data_frame_handler import DataFrameHandler
+from aito.schema import AitoTableSchema
+from aito.common.data_frame_handler import DataFrameHandler
 from .sub_command import SubCommand
 from ..parser import InputArgType, ParseError
 from ..parser_utils import create_sql_connecting_from_parsed_args
